Replace clusterclient_async_sequence with clusterclient_async
The clusterclient_async test binary now provides all features that
previously only existed in clusterclient_async_sequence.
clusterclient_async_sequence is therefor removed.

Inherits clusterclient_async_sequence's behaviour by sending commands
sequentially instead of queuing all commands before starting the event engine.
This allows testcases to perform behavioral changes between commands using
action commands. See clusterclient_async.c.
