+++ tests/CMakeLists.txt
@@ -153,8 +153,6 @@ add_executable(clusterclient_async clusterclient_async.c)
 target_link_libraries(clusterclient_async hiredis_cluster ${SSL_LIBRARY} ${LIBEVENT_LIBRARY})
 add_executable(clusterclient_reconnect_async clusterclient_reconnect_async.c)
 target_link_libraries(clusterclient_reconnect_async hiredis_cluster ${SSL_LIBRARY} ${LIBEVENT_LIBRARY})
-add_executable(clusterclient_async_sequence clusterclient_async_sequence.c)
-target_link_libraries(clusterclient_async_sequence hiredis_cluster ${SSL_LIBRARY} ${LIBEVENT_LIBRARY})
 add_test(NAME set-get-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/set-get-test.sh"
                  "$<TARGET_FILE:clusterclient>"
@@ -193,11 +191,11 @@ add_test(NAME dbsize-to-all-nodes-test
                  WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME dbsize-to-all-nodes-test-async
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/dbsize-to-all-nodes-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
                  WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME dbsize-to-all-nodes-during-scaledown-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/dbsize-to-all-nodes-during-scaledown-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
                  WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME reconnect-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/reconnect-test.sh"
@@ -205,7 +203,7 @@ add_test(NAME reconnect-test
                 WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME timeout-handling-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/timeout-handling-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME connect-error-using-cluster-nodes-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/connect-error-using-cluster-nodes-test.sh"
@@ -213,19 +211,19 @@ add_test(NAME connect-error-using-cluster-nodes-test
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME command-from-callback-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/command-from-callback-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME ask-redirect-connection-error-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/ask-redirect-connection-error-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME cluster-down-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/cluster-down-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME connection-error-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/connection-error-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 add_test(NAME redirect-with-hostname-test
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/redirect-with-hostname-test.sh"
@@ -247,7 +245,7 @@ endif()
 # Disabling the testcase if hiredis contains the issue or if the version is unknown.
 add_test(NAME redirect-with-hostname-test-async
          COMMAND "${CMAKE_SOURCE_DIR}/tests/scripts/redirect-with-hostname-test.sh"
-                 "$<TARGET_FILE:clusterclient_async_sequence>"
+                 "$<TARGET_FILE:clusterclient_async>"
          WORKING_DIRECTORY "${CMAKE_SOURCE_DIR}/tests/scripts/")
 if(hiredis_VERSION VERSION_EQUAL "1.1.0" OR hiredis_VERSION VERSION_EQUAL "0")
   set_tests_properties(redirect-with-hostname-test-async PROPERTIES DISABLED True)
